extract and set patches tests in single file
